+++ app/src/main/java/at/andreasrohner/spartantimelapserec/DeviceStatusReceiver.java
@@ -23,7 +23,6 @@ import android.content.Context;
 import android.content.Intent;
 import android.content.SharedPreferences;
 import android.preference.PreferenceManager;
-import android.util.Log;
 
 public class DeviceStatusReceiver extends BroadcastReceiver {
