--- app/src/main/java/at/andreasrohner/spartantimelapserec/preference/EditSummaryPreference.java
@@ -1,32 +1,0 @@
-package at.andreasrohner.spartantimelapserec.preference;
-
-import android.content.Context;
-import android.preference.EditTextPreference;
-import android.util.AttributeSet;
-
-/**
- * Show value in summary
- */
-public class EditSummaryPreference extends EditTextPreference {
-
-    public EditSummaryPreference(Context context, AttributeSet attrs, int defStyleAttr, int defStyleRes) {
-        super(context, attrs, defStyleAttr, defStyleRes);
-    }
-
-    public EditSummaryPreference(Context context, AttributeSet attrs, int defStyleAttr) {
-        super(context, attrs, defStyleAttr);
-    }
-
-    public EditSummaryPreference(Context context, AttributeSet attrs) {
-        super(context, attrs);
-    }
-
-    public EditSummaryPreference(Context context) {
-        super(context);
-    }
-
-    @Override
-    public CharSequence getSummary() {
-        return getText();
-    }
-}
