--- app/src/main/java/at/andreasrohner/spartantimelapserec/preference/NoKBEditTextPreference.java
@@ -1,59 +1,0 @@
-/*
- * Spartan Time Lapse Recorder - Minimalistic android time lapse recording app
- * Copyright (C) 2014  Andreas Rohner
- *
- * This program is free software: you can redistribute it and/or modify
- * it under the terms of the GNU General Public License as published by
- * the Free Software Foundation, either version 3 of the License, or
- * (at your option) any later version.
- *
- * This program is distributed in the hope that it will be useful,
- * but WITHOUT ANY WARRANTY; without even the implied warranty of
- * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
- * GNU General Public License for more details.
- *
- * You should have received a copy of the GNU General Public License
- * along with this program.  If not, see <http://www.gnu.org/licenses/>.
- */
-
-package at.andreasrohner.spartantimelapserec.preference;
-
-import android.content.Context;
-import android.content.DialogInterface;
-import android.preference.EditTextPreference;
-import android.util.AttributeSet;
-import android.view.View;
-import android.view.inputmethod.InputMethodManager;
-
-public class NoKBEditTextPreference extends EditTextPreference {
-	private View mView;
-
-	@Override
-	protected void onBindView(View view) {
-		mView = view;
-		super.onBindView(view);
-	}
-
-	public NoKBEditTextPreference(Context context) {
-		super(context);
-	}
-
-	public NoKBEditTextPreference(Context context, AttributeSet attrs,
-			int defStyle) {
-		super(context, attrs, defStyle);
-	}
-
-	public NoKBEditTextPreference(Context context, AttributeSet attrs) {
-		super(context, attrs);
-	}
-
-	@Override
-	public void onDismiss(DialogInterface dialog) {
-		super.onDismiss(dialog);
-
-		InputMethodManager imm = (InputMethodManager) getContext()
-				.getSystemService(Context.INPUT_METHOD_SERVICE);
-		imm.hideSoftInputFromWindow(mView.getWindowToken(),
-				InputMethodManager.HIDE_NOT_ALWAYS);
-	}
-}
