+++ app/src/main/java/at/andreasrohner/spartantimelapserec/preference/SeekBarPreference.java
@@ -25,6 +25,7 @@ import android.util.AttributeSet;
 import android.view.View;
 import android.widget.SeekBar;
 import android.widget.TextView;
+
 import at.andreasrohner.spartantimelapserec.R;
 
 public class SeekBarPreference extends DialogPreference implements
