+++ app/src/main/java/at/andreasrohner/spartantimelapserec/recorder/ImageRecorder.java
@@ -18,13 +18,6 @@
 
 package at.andreasrohner.spartantimelapserec.recorder;
 
-import java.io.File;
-import java.io.FileOutputStream;
-import java.io.IOException;
-import java.util.HashSet;
-import java.util.List;
-import java.util.Set;
-
 import android.content.Context;
 import android.graphics.ImageFormat;
 import android.graphics.SurfaceTexture;
@@ -35,6 +28,13 @@ import android.os.Handler;
 import android.os.Looper;
 import android.os.SystemClock;
 import android.util.Log;
+
+import java.io.File;
+import java.io.FileOutputStream;
+import java.io.IOException;
+import java.util.HashSet;
+import java.util.List;
+import java.util.Set;
 
 import at.andreasrohner.spartantimelapserec.data.RecSettings;
