+++ app/src/main/java/at/andreasrohner/spartantimelapserec/settings/RestSettingsActivity.java
@@ -2,13 +2,10 @@ package at.andreasrohner.spartantimelapserec.settings;
 
 import android.content.Context;
 import android.content.SharedPreferences;
-import android.os.Bundle;
-
 import android.widget.Toast;
 
 import androidx.annotation.Nullable;
 import androidx.preference.EditTextPreference;
-import androidx.preference.PreferenceFragmentCompat;
 import at.andreasrohner.spartantimelapserec.R;
 
 /**
