+++ app/src/main/java/at/andreasrohner/spartantimelapserec/settings/ShowActivityPreference.java
@@ -5,12 +5,10 @@ import android.content.Context;
 import android.content.SharedPreferences;
 import android.util.AttributeSet;
 import android.util.Log;
-import android.widget.Toast;
 
 import androidx.annotation.NonNull;
 import androidx.annotation.Nullable;
 import androidx.preference.DialogPreference;
-import androidx.preference.Preference;
 import at.andreasrohner.spartantimelapserec.rest.HttpThread;
 
 /**
