+++ app/src/main/res/drawable/ic_folder.xml
@@ -1,0 +1,31 @@
+<vector xmlns:android="http://schemas.android.com/apk/res/android"
+    android:width="50dp"
+    android:height="50dp"
+    android:viewportWidth="5000"
+    android:viewportHeight="5000">
+  <group>
+    <clip-path
+        android:pathData="M0,0h5000v5000h-5000z"/>
+    <path
+        android:pathData="M451,1049C508,1018 569,995 632,982L632,696C632,662 641,628 659,598 676,568 702,544 732,526 763,509 798,500 833,500L1998,500C2034,500 2068,509 2099,526 2129,544 2155,568 2172,598 2190,628 2199,662 2199,696L2199,971 4247,971C4353,971 4457,998 4549,1049 4640,1101 4716,1175 4769,1265 4822,1354 4850,1456 4850,1559L4850,3913C4850,4016 4822,4117 4769,4207 4716,4296 4640,4371 4549,4422 4457,4474 4353,4501 4247,4501L4246,4501 753,4501C647,4501 543,4474 451,4422 360,4371 284,4296 231,4207 178,4117 150,4016 150,3913L150,3912 150,1559C150,1456 178,1354 231,1265 284,1175 360,1101 451,1049Z"
+        android:strokeLineJoin="round"
+        android:strokeWidth="28.222"
+        android:fillColor="#FFFFFF"
+        android:strokeColor="#00000000"
+        android:fillType="evenOdd"/>
+    <path
+        android:pathData="M451,1049C508,1018 569,995 632,982L632,696C632,662 641,628 659,598 676,568 702,544 732,526 763,509 798,500 833,500L1998,500C2034,500 2068,509 2099,526 2129,544 2155,568 2172,598 2190,628 2199,662 2199,696L2199,971 4247,971C4353,971 4457,998 4549,1049 4640,1101 4716,1175 4769,1265 4822,1354 4850,1456 4850,1559L4850,3913C4850,4016 4822,4117 4769,4207 4716,4296 4640,4371 4549,4422 4457,4474 4353,4501 4247,4501L4246,4501 753,4501C647,4501 543,4474 451,4422 360,4371 284,4296 231,4207 178,4117 150,4016 150,3913L150,3912 150,1559C150,1456 178,1354 231,1265 284,1175 360,1101 451,1049Z"
+        android:strokeLineJoin="round"
+        android:strokeWidth="100"
+        android:fillColor="#00000000"
+        android:strokeColor="#666666"
+        android:fillType="evenOdd"/>
+    <path
+        android:pathData="M200,1500L4800,1500"
+        android:strokeLineJoin="round"
+        android:strokeWidth="100"
+        android:fillColor="#00000000"
+        android:strokeColor="#666666"
+        android:fillType="evenOdd"/>
+  </group>
+</vector>
