--- app/src/main/res/xml/preferences.xml
@@ -1,46 +1,0 @@
-<PreferenceScreen xmlns:android="http://schemas.android.com/apk/res/android"
-    xmlns:app="http://schemas.android.com/apk/res-auto"
-    android:layout_width="fill_parent"
-    android:layout_height="fill_parent">
-
-    <PreferenceCategory android:title="@string/pref_project_settings_category">
-        <ListPreference
-            android:defaultValue="VIDEO_TIME_LAPSE"
-            android:dialogTitle="@string/pref_rec_mode"
-            android:entries="@array/pref_rec_mode_entries"
-            android:entryValues="@array/pref_rec_mode_values"
-            android:key="pref_rec_mode"
-            android:title="@string/pref_rec_mode" />
-
-        <at.andreasrohner.spartantimelapserec.preference.NoKBEditTextPreference
-            android:defaultValue="@string/app_name"
-            android:inputType="textNoSuggestions"
-            android:key="pref_project_title"
-            android:summary="@string/pref_project_title_summ"
-            android:title="@string/pref_project_title" />
-
-    </PreferenceCategory>
-
-    <PreferenceCategory android:title="@string/pref_camera_settings_category">
-        <at.andreasrohner.spartantimelapserec.preference.CameraInformation
-            android:key="pref_camera_settings_information"
-            android:summary="Loading..."
-            android:title="@string/pref_camera_settings_category" />
-    </PreferenceCategory>
-
-    <PreferenceCategory android:title="@string/pref_scheduling_settings_category">
-        <at.andreasrohner.spartantimelapserec.preference.SchedulingInformation
-            android:key="pref_camera_settings_information"
-            android:summary="Loading..."
-            android:title="@string/pref_scheduling_settings_title" />
-    </PreferenceCategory>
-
-    <PreferenceCategory android:title="@string/pref_stop_settings_category">
-        <at.andreasrohner.spartantimelapserec.preference.StopInformation
-            android:key="pref_stop_settings_information"
-            android:summary="Loading..."
-            android:title="@string/pref_stop_settings_category" />
-    </PreferenceCategory>
-
-
-</PreferenceScreen>
